refactor: 3페이지 보일러플레이트 훅 추출 (#376)
* docs: 전체 문서 최신화 및 최적화 — 코드 간결화 리팩터 반영


* refactor: UserInfoForm 분해 — useUserInfoForm 훅 + BirthTimeInput 추출


* refactor: 3페이지 보일러플레이트 훅 추출 — useResetScrollOnStep, usePreselectCharacter


* fix: sonar.coverage.exclusions에 useUserInfoForm.ts 추가

---------
