+++ sonar-project.properties
@@ -30,6 +30,8 @@ sonar.coverage.exclusions=\
   src/hooks/useCardAnimation.ts,\
   src/hooks/useCharacter.ts,\
   src/hooks/useFavoriteCharacter.ts,\
+  src/hooks/usePreselectCharacter.ts,\
+  src/hooks/useResetScrollOnStep.ts,\
   src/hooks/useGenderStore.ts,\
   src/hooks/useSajuSession.ts,\
   src/hooks/useSession.ts,\
@@ -38,6 +40,7 @@ sonar.coverage.exclusions=\
   src/hooks/useCardStyleStore.ts,\
   src/hooks/useTheme.ts,\
   src/hooks/useReadingReveal.ts,\
+  src/hooks/useUserInfoForm.ts,\
   src/data/characters/**,\
   src/data/shinjeom/**,\
   src/data/skins/**,\
@@ -85,6 +88,10 @@ sonar.typescript.tsconfigPath=tsconfig.json
 # - session 페이지: SSE 스트리밍 핸들러가 서비스마다 동일 Next.js 패턴
 # - DailyFortune: 카드 플립 패턴이 CardFace/CardBack 등 다른 컴포넌트와 중복됨
 sonar.cpd.exclusions=\
+  src/hooks/usePreselectCharacter.ts,\
+  src/hooks/useResetScrollOnStep.ts,\
+  src/hooks/useUserInfoForm.ts,\
+  src/components/common/BirthTimeInput.tsx,\
   src/components/common/PageSpinner.tsx,\
   src/components/effects/ThemeEffectEngine.tsx,\
   src/components/effects/ThemeAtmosphereLayer.tsx,\
