+++ src/app/saju/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useState, useEffect, Suspense } from "react";
+import { useState, Suspense } from "react";
 import { useRouter, useSearchParams } from "next/navigation";
 import { Icon } from "@/components/common/Icon";
 import { motion, AnimatePresence } from "framer-motion";
@@ -17,13 +17,14 @@ import { useGenderStore } from "@/hooks/useGenderStore";
 import { ChatMessage, SajuTimeRange, Topic } from "@/types/session";
 import { UserInfo } from "@/types/user-info";
 import { sajuTimeOptions, sajuAreaOptions, getSajuTimeLabel, getSajuTimeDesc, getSajuAreaLabel, getSajuAreaDesc } from "@/data/saju/categories";
-import { useFavoriteCharacter } from "@/hooks/useFavoriteCharacter";
 import { useLocaleStore } from "@/hooks/useLocaleStore";
 import { ServiceBackground } from "@/components/effects/ServiceBackground";
 import { useT } from "@/i18n/useT";
 import { t as translate } from "@/i18n/translations";
 import type { Locale } from "@/i18n/config";
 import { PageSpinner } from "@/components/common/PageSpinner";
+import { useResetScrollOnStep } from "@/hooks/useResetScrollOnStep";
+import { usePreselectCharacter } from "@/hooks/usePreselectCharacter";
 
 /** "{name}" placeholder 치환 — saju.page.after-info-msg 전용 */
 function buildAfterInfoMsg(name: string, locale: Locale): string {
@@ -230,6 +231,7 @@ function SajuPageContent() {
   const { genderFilter, setGenderFilter } = useGenderStore();
   const sajuCharacters = getCharactersByGender(genderFilter);
 
+  // URL 파라미터 기반 동기 초기화 (flash 방지 — useSearchParams는 Suspense로 래핑된 상태)
   const preselectedCharId = searchParams.get("character");
   const preselectedChar = preselectedCharId ? getCharacterById(preselectedCharId) ?? null : null;
 
@@ -241,35 +243,19 @@ function SajuPageContent() {
   const [monthlyToggle, setMonthlyToggle] = useState(false);
   const [freeQuestion, setFreeQuestionLocal] = useState("");
 
-  // 프리셀렉트된 캐릭터: 스토어 반영 + 인사 메시지 생성 (클라이언트 마운트 후 — new Date() SSR 비결정 방지)
-  useEffect(() => {
-    if (preselectedChar) {
-      setCharacterId(preselectedChar.id);
-      setDialogueMessages([{ id: crypto.randomUUID(), role: "character" as const, content: getCharacterGreeting(preselectedChar, useLocaleStore.getState().locale), mood: "smile", timestamp: new Date() }]);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  // 선호 상담사 fallback: URL 파라미터 없이 직접 접속한 경우 자동 선택
-  const { favoriteCharacter } = useFavoriteCharacter(!!preselectedChar);
-  useEffect(() => {
-    if (favoriteCharacter && !selectedCharacter) {
-      setSelectedCharacter(favoriteCharacter);
-      setCharacterId(favoriteCharacter.id);
-      setDialogueMessages([{ id: crypto.randomUUID(), role: "character", content: getCharacterGreeting(favoriteCharacter, useLocaleStore.getState().locale), mood: "smile", timestamp: new Date() }]);
+  // 프리셀렉트 + 즐겨찾기 fallback: 캐릭터 선택 시 스토어 반영 + 인사 메시지 생성 + 스텝 전환
+  usePreselectCharacter({
+    currentCharacter: selectedCharacter,
+    onSelect: (character) => {
+      setSelectedCharacter(character);
+      setCharacterId(character.id);
+      setDialogueMessages([{ id: crypto.randomUUID(), role: "character" as const, content: getCharacterGreeting(character, useLocaleStore.getState().locale), mood: "smile", timestamp: new Date() }]);
       setStep("info-input");
-    }
-  }, [favoriteCharacter, selectedCharacter, setCharacterId]);
+    },
+  });
 
-  // 스텝 전환 시 스크롤 최상단 초기화 (3중 보정: 즉시 + rAF + rAF)
-  useEffect(() => {
-    const resetScroll = () => {
-      window.scrollTo({ top: 0, left: 0, behavior: "instant" });
-      document.querySelectorAll("[class*='overflow-y-auto'], [class*='overflow-auto']").forEach((el) => { el.scrollTop = 0; });
-    };
-    resetScroll();
-    requestAnimationFrame(() => { resetScroll(); requestAnimationFrame(resetScroll); });
-  }, [step]);
+  // 스텝 전환 시 스크롤 최상단 초기화
+  useResetScrollOnStep(step);
 
   const handleCharacterSelect = (character: CharacterConfig) => {
     setSelectedCharacter(character);
